Add a template for new services
To reduce drift between service we need to generate and maintain them in
a shared way where possible.

This is first just a simple template that can be used as a basis.

But we should look into tools that can continue to maintain and share
these files where desired after the initial setup